Upgrade discord.py to 1.5.1 (#43473)
* Upgrade discord.py to 1.5.1

discord.py 1.5.1 introduces support for Intents that are required by future bots and integrations. While this is not yet required by Home Assistant Discord integration it is needed in the future, because old API which discord.py 1.4.1 uses is now deprecated. Also it would solve dependency collision with my custom component discord_game.

* Upgrade discord.py to 1.5.1